+++ semantic_admin/contrib/import_export/widgets.py
@@ -1,7 +1,0 @@
-from semantic_forms.widgets import SemanticSelect
-
-
-class SemanticExportActionSelect(SemanticSelect):
-    """Semantic export action select."""
-
-    template_name = "admin/import_export/export_action_select.html"