+++ semantic_admin/tests/test_templatetags.py
@@ -1,6 +1,11 @@
 from django.test import SimpleTestCase, override_settings
+from django.utils.safestring import mark_safe
 
 from semantic_admin.templatetags.semantic_app_list import get_semantic_app_list
+from semantic_admin.templatetags.semantic_filters import (
+    semantic_error_list,
+    semantic_help_text,
+)
 
 
 def app_list():
@@ -59,3 +64,16 @@ class SemanticAppListTests(SimpleTestCase):
 
         self.assertTrue(apps[0]["is_active"])
 
+
+class SemanticFormRenderTests(SimpleTestCase):
+    def test_semantic_help_text_adds_semantic_list_classes(self):
+        html = semantic_help_text(mark_safe("<ul><li>One</li><li>Two</li></ul>"))
+
+        self.assertIn('<ul class="ui bulleted list">', html)
+        self.assertEqual(html.count('<li class="item">'), 2)
+
+    def test_semantic_error_list_uses_semantic_list_classes(self):
+        html = semantic_error_list(["One", "Two"])
+
+        self.assertIn('<ul class="ui bulleted list semantic-error-list">', html)
+        self.assertEqual(html.count('<li class="item">'), 2)