+++ docs/changes.md
@@ -1,3 +1,8 @@
+v0.6.9
+------
+* Added semantic user and group admins
+* Removed django-import-export integration
+
 v0.6.8
 ------
 * Upgrade Fomantic v2.9.4 and use Django admin jQuery